Improve mock reliability
`patch()` calls at the `setUp` level only affect the calls in setUp
(which is where the charm is initialized), but not the calls in `test_`
methods.

`patch()` calls at the `TestCharm` level only affect the `test_`
methods, and do not affect charm initialization.

This caused some incosistency and unreliability of the mocks, depending
on when certain calls were made, and on multiple occasions the tests
attempted to install snaps on my machine. By mocking the entire class
(instead of just individual methods), and using a patcher with
`start()`/`stop()` calls in `setUp()` and `tearDown()`, we are patching
all the calls made in the test, and avoid letting calls slip through the
cracks. It also avoids having patch calls at multiple different levels.